Support multiple delegates via io.modelcontextprotocol.server.Delegates
